+++ CHANGES.rst
@@ -5,7 +5,10 @@ Changelog
 1.3.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fixed applying the demo profile at new Plone Site creation time.
+  This was due to BrowserLayers still not initialized, in this case we mark the
+  `REQUEST` with registred `BrowserLayers` ourselves.
+  [gbastien]
 
 
 1.3.3 (2021-06-28)